phase4_api/requirements.txt
fastapi[all]
uvicorn
pydantic
python-dotenv
aiofiles
groq
google-generativeai
